requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
spotipy>=2.24.0
ytmusicapi>=1.8.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
itsdangerous>=2.1.0
thefuzz>=0.22.0
python-Levenshtein>=0.25.0
sse-starlette>=2.0.0
